requirements.txt
argcomplete==1.10.3
attrs==22.1.0
beautifulsoup4==4.8.2
chardet==3.0.4
click==8.1.3
compressed-rtf==1.0.6
distutils-pytest==0.2
docx2txt==0.8
ebcdic==1.1.1
et-xmlfile==1.1.0
extract-msg==0.28.7
Flask==2.2.2
Flask-Cors==3.0.10
gunicorn==20.1.0
IMAPClient==2.1.0
iniconfig==1.1.1
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.1
MarkupSafe==2.1.1
numpy==1.23.4
olefile==0.46
openpyxl==3.0.10
packaging==21.3
pandas==1.5.1
pdfminer.six==20191110
Pillow==9.2.0
pluggy==1.0.0
py==1.11.0
pycryptodome==3.15.0
pyparsing==3.0.9
PyPDF2==2.11.0
pytest==7.1.3
python-dateutil==2.8.2
python-docx==0.8.11
python-pptx==0.6.21
pytz==2022.5
pytz-deprecation-shim==0.1.0.post0
six==1.12.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SpeechRecognition==3.8.1
textract==1.6.5
tomli==2.0.1
tzdata==2022.4
tzlocal==4.2
Werkzeug==2.2.2
xlrd==1.2.0
XlsxWriter==3.0.3
